packages/phantom-intake/requirements.txt
requests>=2.28.1
beautifulsoup4>=4.11.1
rich>=12.0.0
pyyaml>=6.0